+++ src/main/java/com/kernel/sense_log/domain/service/DiaryLikeService.java
@@ -1,7 +1,7 @@
 package com.kernel.sense_log.domain.service;
 
 import com.kernel.sense_log.domain.entity.enumeration.Tag;
-import com.kernel.sense_log.web.dto.DiaryLikeResDTO;
+import com.kernel.sense_log.web.dto.response.DiaryLikeResDTO;
 
 public interface DiaryLikeService {
     DiaryLikeResDTO saveOrUpdateEmotion(Long userId, Long diaryId, Tag emotion);
