+++ src/main/java/com/kernel/sense_log/domain/service/impl/SubTagServiceImpl.java
@@ -1,7 +1,8 @@
-package com.kernel.sense_log.domain.service;
+package com.kernel.sense_log.domain.service.impl;
 
 import com.kernel.sense_log.domain.entity.SubTag;
 import com.kernel.sense_log.domain.repository.SubTagRepository;
+import com.kernel.sense_log.domain.service.SubTagService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -9,7 +10,7 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class SubTagServiceImpl implements SubTagService{
+public class SubTagServiceImpl implements SubTagService {
 
     private final SubTagRepository subTagRepository;
     @Override
