+++ src/main/java/com/kernel/sense_log/web/controller/AuthController.java
@@ -5,10 +5,10 @@ import com.kernel.sense_log.common.dto.ResultObject;
 import com.kernel.sense_log.domain.auth.jwt.JwtUtil;
 import com.kernel.sense_log.domain.entity.User;
 import com.kernel.sense_log.domain.service.AuthService;
-import com.kernel.sense_log.domain.service.UserServiceImpl;
-import com.kernel.sense_log.web.dto.LoginResDTO;
-import com.kernel.sense_log.web.dto.UserReqDTO;
-import com.kernel.sense_log.web.dto.UserResDTO;
+import com.kernel.sense_log.domain.service.impl.UserServiceImpl;
+import com.kernel.sense_log.web.dto.response.LoginResDTO;
+import com.kernel.sense_log.web.dto.request.UserReqDTO;
+import com.kernel.sense_log.web.dto.response.UserResDTO;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
