+++ src/main/java/com/kernel/sense_log/web/controller/DiaryLikeController.java
@@ -1,7 +1,7 @@
 package com.kernel.sense_log.web.controller;
 
 import com.kernel.sense_log.domain.entity.enumeration.Tag;
-import com.kernel.sense_log.web.dto.DiaryLikeResDTO;
+import com.kernel.sense_log.web.dto.response.DiaryLikeResDTO;
 import com.kernel.sense_log.domain.service.DiaryLikeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
