+++ src/main/java/com/kernel/sense_log/web/dto/request/UserReqDTO.java
@@ -1,4 +1,4 @@
-package com.kernel.sense_log.web.dto;
+package com.kernel.sense_log.web.dto.request;
 
 import com.kernel.sense_log.domain.entity.User;
 import jakarta.persistence.Column;
